Begin switch of pet information to inventory.

+++ atlas.com/channel/kafka/consumer/message/consumer.go
@@ -186,7 +186,7 @@ func handlePetChat(sc server.Model, wp writer.Producer) message.Handler[message3
 			return
 		}
 
-		p := pet.NewModelBuilder(uint64(e.ActorId), 0, 0, "").SetOwnerID(e.Body.OwnerId).SetSlot(e.Body.PetSlot).Build()
+		p := pet.NewModelBuilder(e.ActorId, 0, 0, "").SetOwnerID(e.Body.OwnerId).SetSlot(e.Body.PetSlot).Build()
 		_ = _map.NewProcessor(l, ctx).ForSessionsInMap(s.Map(), session.Announce(l)(ctx)(wp)(writer.PetChat)(writer.PetChatBody(p, e.Body.Type, e.Body.Action, e.Message, e.Body.Balloon)))
 	}
 }
