+++ atlas.com/channel/socket/writer/pet_exclude.go
@@ -11,7 +11,7 @@ func PetExcludeResponseBody(p pet.Model) BodyProducer {
 	return func(w *response.Writer, options map[string]interface{}) []byte {
 		w.WriteInt(p.OwnerId())
 		w.WriteInt8(p.Slot())
-		w.WriteLong(p.Id())
+		w.WriteLong(uint64(p.Id()))
 		w.WriteByte(byte(len(p.Excludes())))
 		for _, e := range p.Excludes() {
 			w.WriteInt(e.ItemId())
